backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
numpy==1.26.4
aiocache==0.11.1